Add .travis.yml and tox.ini for automated testing
This uses tox for setting up the environment and test-runner. The
dependencies for this are added to setup.py if you want to run pytest
manually with your own virtual environment, while tox is used for
automatically setting up the environment.

This will run tests with dependencies on spark on travis, and then send
the results to codecov.io.