Bug 1357875 - Use collect instead of writing directly to `file://` via spark
There are certain permission issues when deploying topline_dashboard due to permission errors with the spark executor. One alternative is to write the csv files to hdfs and copy over to local disk. This is not necessarily an elegant solution to the problem, so the tried solution is to use the collect (as long as this dataset stays within the size of the driver).