Bug 1357875 - Use toLocalIterator instead of `file://`
This also:
* Fixes dataframe collect methods under mock_s3 context
* Adds tests for utils.write_csv

There are certain permission issues when deploying topline_dashboard due
to permission errors with the spark executor. One alternative is to
write the csv files to hdfs and copy over to local disk. This is not
necessarily an elegant solution to the problem, so the tried solution is
to use the collect (as long as this dataset stays within the size of the
driver).

This also takes special care of unicode encoded data, and adds a test
case for both ascii and utf-8.

When running methods like df.collect or df.toLocalIterator under the
mock_s3 context, the result will always be an empty list. [1] The
workaround is to mock the call to toLocalIterator and the call to
write_csv, depending on the context of the test.

[1] https://github.com/spulec/moto/issues/929