Bug 1364530 - Migrate Search ETL
This refactors the search rollup to handle both daily and monthly search
rollups. The output schema should be comparable to the previous version
of the rollup.

In addition to functional parity, adds testing for general logic
and the supporting infrastructure code (cli, manifest generation). The
job is also airflow friendly.