Bug 1389231 - Add new-profile ping to churn
This adds the `new-profile` ping as a data source for churn.

This is also a version bump to `v3`.

Additionally, `spark.sql.session.timeZone` is set to UTC, which require
Spark 2.2.0 to work (see SPARK-18936 for details). The spark session
assumes the timezone of the local machine, which can cause failing
tests.