Bug 1381840 - Add 1 day Firefox retention intermediate dataset
In it's current state, this should be considered `retention_dev_v3`. The
link to the STMO example dashboard for related datasets can be found in
[1].

This contains the cleaning routines for `main_summary` and `new_profile`
for use in a 1 day retention day dataset. There can be several instances
of the same client across a day (on average ~3 pings per client). This
is later processed by aggregating using HyperLogLog UDF's in
Telemetry-Batch-View.

This contains necessary columns for calculating retention by
attribution over profile counts and for calculating engagement over
subpopulations.

The extract routine is borrowed from churn, while transform uses the
same subroutines for cleaning and transforming the dataset.

[1]
https://sql.telemetry.mozilla.org/dashboard/firefox-telemetry-retention-dataset-example-usage