bug 1775444: drop top signatures from 250 to 200
We're having memory issues and it looks like the total number of crash
reports for the top signatures is increasing over time. This drops the
number of top signatures we look at which will drop the number of crash
reports which should reduce the memory used back to a workable number.